backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-filter==23.1
djoser==2.1.0
Pillow==9.0.0
unidecode
short_url
gunicorn==20.1.0
python-dotenv
psycopg2-binary==2.9.3 
